#7 オニオンアーキテクチャのファイル構成に変更
